+++ HKLMart/src/main/java/com/hklmart/controller/BoardController.java
@@ -44,9 +44,5 @@ public class BoardController {
         return "redirect:/member/my-page";
     }
 
-    @PostMapping("/ask-answer-up")
-    public String askAnswerUp(BoardAskVO boardAskVO) {
-        boardService.updateAskAnswer(boardAskVO);
-        return "redirect:/manager/manager-ask-board";
-    }
+
 }
