+++ HKLMart/src/main/java/com/hklmart/persistence/BoardDAO.java
@@ -9,6 +9,6 @@ public interface BoardDAO {
 
     void insertBoardReview(BoardReviewVO boardReviewVO);
 
-    void updateAskAnswer(BoardAskVO boardAskVO);
+
 
 }
