+++ HKLMart/src/main/java/com/hklmart/service/BoardService.java
@@ -82,9 +82,7 @@ public class BoardService {
         return path;
     }
 
-    public void updateAskAnswer(BoardAskVO boardAskVO) {
-        boardDAO.updateAskAnswer(boardAskVO);
-    }
+
 
 
 }
