+++ HKLMart/src/main/resources/com/hklmart/persistence/ManagementDAO.xml
@@ -191,4 +191,13 @@
         WHERE ORDER_NUMBER = #{orderNumber}
 
     </update>
+
+    <update id="updateAskAnswer">
+        <![CDATA[
+        update BOARD_ASK
+        set BOARD_ASK_STATE  = #{boardAskState},
+            BOARD_ASK_ANSWER = #{boardAskAnswer}
+        where BOARD_ASK_NUMBER = #{boardAskNumber}
+        ]]>
+    </update>
 </mapper>
