+++ HKLMart/src/main/webapp/resources/css/product-detail-all.css
@@ -23908,7 +23908,7 @@ a:link {
 
 #header {
     position: relative;
-    z-index: 21;
+    z-index: 510;
 }
 
 #container {
